pyproject.toml
[project]
name = "aissert"
version = "0.0.1-alpha"
description = "Library for metrics and tests on NLP sets"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "decorator",
    "click",
    # "annotated-types==0.7.0",
    # "anyio==4.8.0",
    # "certifi==2025.1.31",
    # "charset-normalizer==3.4.1",
    # "click==8.1.8",
    # "fastapi==0.115.8",
    # "filelock==3.17.0",
    # "fsspec==2025.2.0",
    # "h11==0.14.0",
    # "huggingface-hub==0.28.1",
    # "idna==3.10",
    # "Jinja2==3.1.5",
    # "MarkupSafe==3.0.2",
    # "mpmath==1.3.0",
    # "networkx==3.4.2",
    # "numpy==2.2.3",
    # "packaging==24.2",
    # "pydantic==2.10.6",
    # "pydantic_core==2.27.2",
    "PyYAML==6.0.2",
    # "regex==2024.11.6",
    # "requests==2.32.3",
    # "safetensors==0.5.2",
    # "setuptools==75.8.0",
    # "sniffio==1.3.1",
    # "starlette==0.45.3",
    # "sympy==1.13.1",
    # "tokenizers==0.21.0",
    # "torch==2.6.0",
    # "tqdm==4.67.1",
    # "transformers==4.48.3",
    # "typing_extensions==4.12.2",
    # "urllib3==2.3.0",
    # "uvicorn==0.34.0",
    "langid",
]

[dependency-groups]
dev = [
    "pytest>=8.3.3",
    # "debugpy>=1.8.11"
]

